Рефакторинг модуля utils.js
