+++ src/mock/generateEvent.js
@@ -1,4 +1,4 @@
-import {getRandomInteger, getRandomArrayElement} from "../utils.js";
+import {getRandomInteger, getRandomArrayElement} from "../utils/common.js";
 import {
   DESCRIPTIONS,
   EVENT_TYPES,
