+++ src/utils/render.js
@@ -1,0 +1,55 @@
+import Abstract from "../view/Abstract.js";
+
+const renderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  AFTEREND: `afterend`,
+  BEFOREBEGIN: `beforebegin`,
+  BEFOREEND: `beforeend`,
+};
+
+const render = (container, element, position) => {
+  if (container instanceof Abstract) {
+    container = container.getElement();
+  }
+  if (element instanceof Abstract) {
+    element = element.getElement();
+  }
+  switch (position) {
+    case renderPosition.BEFOREBEGIN:
+      container.before(element);
+      break;
+    case renderPosition.BEFOREEND:
+      container.append(element);
+      break;
+    case renderPosition.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case renderPosition.AFTEREND:
+      container.after(element);
+      break;
+  }
+};
+
+const replace = (newChild, oldChild) => {
+  if (oldChild instanceof Abstract) {
+    oldChild = oldChild.getElement();
+  }
+  if (newChild instanceof Abstract) {
+    newChild = newChild.getElement();
+  }
+  const parent = oldChild.parentElement;
+  if (parent === null || oldChild === null || newChild === null) {
+    throw new Error(`Can't replace unexisting elements`);
+  }
+  parent.replaceChild(newChild, oldChild);
+};
+
+const remove = (component) => {
+  if (!(component instanceof Abstract)) {
+    throw new Error(`Can remove only components`);
+  }
+  component.getElement().remove();
+  component.removeElement();
+};
+
+export {renderPosition, render, replace, remove};
