+++ src/view/Abstract.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import {createElement} from "../utils/events.js";
 
 export default class Abstract {
   constructor() {
